backend/requirements.txt
Django==6.0
django-cors-headers==4.6.0

djangorestframework==3.15.2

psycopg2-binary==2.9.10

python-dotenv==1.0.1

pypandoc==1.15
pylatexenc==2.10

Pillow==11.0.0

python-magic==0.4.27
bleach==6.1.0